requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
python-pptx>=0.6.21
moviepy>=1.0.3
pywin32>=305
matplotlib>=3.7.0
